backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
yfinance==0.2.54
groq==0.9.0
httpx<0.28
python-dotenv==1.0.1
pandas==2.2.2
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
